+++ platform/libretro/msvc/msvc-2010.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "msvc-2010", "msvc-2010\msvc-2010.vcxproj", "{D4156C25-0E30-4407-9198-1F51EF74AA84}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Release|Win32 = Release|Win32
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D4156C25-0E30-4407-9198-1F51EF74AA84}.Debug|Win32.ActiveCfg = Debug|Win32
+		{D4156C25-0E30-4407-9198-1F51EF74AA84}.Debug|Win32.Build.0 = Debug|Win32
+		{D4156C25-0E30-4407-9198-1F51EF74AA84}.Release|Win32.ActiveCfg = Release|Win32
+		{D4156C25-0E30-4407-9198-1F51EF74AA84}.Release|Win32.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
